+++ packages/test-dialog-pane/src/tests-react.js
@@ -1,9 +1,14 @@
 import React, { Component } from "react"; // eslint-disable-line no-unused-vars
 import { renderer, keys, DialogPane, Toolbar, ToolbarTitle, Button } from "polythene-react";
 import genericTests from "./tests-generic";
-import { longText, cancelOkButtons } from "./shared";
+import { longText } from "./shared";
 
 const reactTests = () => {
+
+  const cancelOkButtons = [
+    <Button label="Cancel" key="cancel" />,
+    <Button label="Save" key="save" />,
+  ];
 
   return [
     {
@@ -15,7 +20,7 @@ const reactTests = () => {
         <DialogPane
           title="Long dialog with a very long title that surely won't fit here"
           body={renderer.trust(longText)}
-          footer={cancelOkButtons({ renderer, Button })}
+          footerButtons={cancelOkButtons}
         />
     },
   ];